feat: set service_name from Log Group and hostname from Log Stream suffix
This sets the New Relic Logs attribute `service_name` to the CloudWatch
Log Group, and sets the attribute `hostname` to the trailing portion of
the CloudWatch Log Stream ("/" delimited, the ECS task ID for ECS logs).
Without this change, both of those fields are N/A in NR Logs if the
application is not setting them in the log message.